+++ backend/modules/controllers/exam-result-structure.js
@@ -5,9 +5,10 @@ const NotificationModel = require('../models/notification');
 const classModel = require('../models/class');
 
 let GetSingleClassExamResultStructure = async (req, res, next) => {
-    let className = req.params.id;
+    let adminId = req.params.id;
+    let className = req.params.class;
     try {
-        const singleExamResultStructureStr = await ExamResultStructureModel.find({ class: className });
+        const singleExamResultStructureStr = await ExamResultStructureModel.find({ adminId: adminId, class: className });
         return res.status(200).json(singleExamResultStructureStr);
     } catch (error) {
         return res.status(500).json('Internal Server Error !');;
@@ -15,6 +16,7 @@ let GetSingleClassExamResultStructure = async (req, res, next) => {
 }
 
 let GetSingleClassExamResultStructureByStream = async (req, res, next) => {
+    let adminId = req.params.id;
     let className = req.params.class;
     let stream = req.params.stream;
     let examType = req.params.exam;
@@ -23,7 +25,7 @@ let GetSingleClassExamResultStructureByStream = async (req, res, next) => {
     }
     let streamMsg = `${stream} stream`;
     try {
-        const singleExamResultStructureStr = await ExamResultStructureModel.findOne({ class: className, stream: stream, examType: examType });
+        const singleExamResultStructureStr = await ExamResultStructureModel.findOne({adminId:adminId, class: className, stream: stream, examType: examType });
         if (!singleExamResultStructureStr) {
             if (stream === "N/A") {
 
@@ -39,17 +41,18 @@ let GetSingleClassExamResultStructureByStream = async (req, res, next) => {
 
 let CreateExamResultStructure = async (req, res, next) => {
     let className = req.body.class;
-    let { examType, stream } = req.body;
+    let { adminId, examType, stream } = req.body;
     let { theoryMaxMarks, theoryPassMarks, practicalMaxMarks, practicalPassMarks, gradeMinMarks, gradeMaxMarks } = req.body.type;
     if (stream === "stream") {
         stream = "N/A";
     }
     try {
-        const checkExamExist = await ExamResultStructureModel.findOne({ class: className, examType: examType, stream: stream });
+        const checkExamExist = await ExamResultStructureModel.findOne({ adminId: adminId, class: className, examType: examType, stream: stream });
         if (checkExamExist) {
             return res.status(400).json(`This class ${examType} exam structure already exist !`);
         }
         let examResultStructureData = {
+            adminId:adminId,
             class: className,
             examType: examType,
             stream: stream,
@@ -78,14 +81,14 @@ let DeleteResultStructure = async (req, res, next) => {
     try {
         const id = req.params.id;
         const resultStr = await ExamResultStructureModel.findOne({ _id: id });
-        const { class: className, stream, examType } = resultStr;
+        const {adminId, class: className, stream, examType } = resultStr;
         const deleteResultStructure = await ExamResultStructureModel.findByIdAndRemove(id);
         if (deleteResultStructure) {
-            const result = await ExamResultModel.findOne({ class: className, stream: stream, examType: examType });
+            const result = await ExamResultModel.findOne({adminId:adminId, class: className, stream: stream, examType: examType });
             if (!result) {
-                return res.status(200).json('Exam Result structure delete successfully.');
+                return res.status(200).json('Exam Result structure not found.');
             }
-            const deleteResult = await ExamResultModel.deleteMany({ class: className, stream: stream, examType: examType });
+            const deleteResult = await ExamResultModel.deleteMany({adminId:adminId, class: className, stream: stream, examType: examType });
             if (deleteResult) {
                 return res.status(200).json('Exam Result structure delete successfully.');
             }
@@ -101,32 +104,33 @@ let ChangeResultPublishStatus = async (req, res, next) => {
         if (!examResultStr) {
             return res.status(200).json('Exam result structure not found !');
         }
-        const findResultPublishStatus = await examResultStr.resultPublishStatus;
-        const cls = await examResultStr.class;
-        const stream = await examResultStr.stream;
-        const examType = await examResultStr.examType;
+        const findResultPublishStatus = examResultStr.resultPublishStatus;
+        const adminId = examResultStr.adminId;
+        const cls = examResultStr.class; 
+        const stream = examResultStr.stream;
+        const examType = examResultStr.examType;
 
-        const isExamResultExist = await ExamResultModel.findOne({ class: cls, stream: stream, examType: examType });
-        if(!isExamResultExist && findResultPublishStatus==false){
+        const isExamResultExist = await ExamResultModel.findOne({adminId:adminId, class: cls, stream: stream, examType: examType });
+        if (!isExamResultExist && findResultPublishStatus == false) {
             return res.status(404).json('Exam result not found !');
         }
-        let title='';
-        let message ='';
+        let title = '';
+        let message = '';
         if (findResultPublishStatus == false) {
             let className;
-            if(cls==1){
+            if (cls == 1) {
                 className = `${cls}st`
             }
-            if(cls==2){
+            if (cls == 2) {
                 className = `${cls}nd`
             }
-            if(cls==3){
+            if (cls == 3) {
                 className = `${cls}rd`
             }
-            if(cls >= 4 && cls <= 12){
+            if (cls >= 4 && cls <= 12) {
                 className = `${cls}th`
             }
-            if(cls==200){
+            if (cls == 200) {
                 className = `Nursery`;
             }
             if (cls == 201) {
@@ -145,7 +149,7 @@ let ChangeResultPublishStatus = async (req, res, next) => {
         const updateStatus = await ExamResultStructureModel.findByIdAndUpdate(id, { $set: resultPublishData }, { new: true });
         if (updateStatus) {
             const notification = await NotificationModel.findOne({ class: cls, title: title });
-            if (!notification && title!=='') {
+            if (!notification && title !== '') {
                 const notificationData = {
                     title: title,
                     message: message,
